Allow NULL for vote.vote_type
The get_or_create line was throwing an "Integrity error: vote_type
cannot be NULL", which I think makes sense since we couldn't get a vote
that didn't already have a vote_type set, and we couldn't create a vote
unless we specify the vote_type at the time of creation. I don't know if
it's a bad solution but I changed models.py to allow vote.vote_type to
be NULL, and voting seems to be working properly.

The offending line that prompted me to allow NULL was #337:
vote, created = Vote.objects.get_or_create(voter=v, quote=q)