Eww tech debt; kluge from youtube.js
In order to pass the {{ forloop.counter }} to youtube.js, the
quotes.html must run a forloop over a list. The quote profile page was
receiving only a quote object, so {{ forloop.counter }} was not passing
any number, which resulted in youtube.js reporting null for all objects
and the play button breaking. I 'fixed' the play button by passing both
an object and a list (containing one object) in the context to
quote.html.

After youtube.js is fixed, we can delete the junk on quote.html and
quote.py.