Add time_since_created to Quote
-I imagine there is a library that provides a more efficient solution
than this, but I couldn't find it, and this seems to work for the
meantime.

-This seems to work for me, but I am unclear on how timezones are
handled. I think that auto_add_now is grabbing UTC, so this should work
equally for people in any timezone, but I'm not sure.

-I imagine this piece, if we kept it, should at least be put in a
separate file and imported into models.py, because it could be used for
any class, not just Quote. I'm not sure what a good practice for
implementing this would be.