Add humanize, boolean to Quote model, allow Null on time_quote_ends
-Please refer to issue #58 for the deets on allowing Null. If this is
acceptable we can close issue #58.

-Note change made to settings.py.skel

-Note the schemamigration

-The Boolean podcast_managed_by_superuser will help me design a UI that
shows me (a superuser) all and only the shows in the system that I am
managing, and only the episodes for those podcasts which do not have
YouTube links. With this boolean I can perform a query to populate this
superuser UI with only the episodes I need to add a YouTube video link
to.